Ugly fix for python module rpath issues
The rpath of the python module .so file now links to $DQCSIM_HOME/lib
(evaluated at compile time), which *should* point to ~/.dqcsim/lib,
which *should* contain libdqcshost.so. It'd be much nicer to have this
be a relative link that stays within the egg directory somehow, but holy
shit is ld giving me a hard time with that. The best solution so far
seems to be to compile libdqcshost.so as a static library instead of
cdylib, and then post-build it to a shared object. Then we can link the
python module with it statically. We're probably going to run into the
exact same issues in other places to though, so eh, we may need rpath
nonsense anyway.