Moved the API layer to a more suitable directory
It occurred to me that there will probably be quite a bit of overlap
between the host and plugin APIs (particularly the arb part) which would
be hard not to duplicate if the host and plugin APIs were different crates.
The Python module is similarly annoying if we want to have a single base
module named "dqcsim". So I just put them together. At that point it
also makes sense to just stick it in the root along with the rest of the
rust crates. The "foreign" dir may return for C/C++/Python testing
logic though.