Add raw closure mode for PluginThreadConfiguration
That is, allow the closure that actually makes the PluginState to be
user-defined, as an alternative to specifying the PluginDefinition. This
allows all API functions to be tested, since the entire contents of a
plugin's main() can be wrapped inside this raw closure (i.e.,
dqcs_plugin_run() or dqcs_plugin_start()).